requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
httpx==0.28.1
chromadb==1.0.15
PyMuPDF==1.26.3
python-docx==1.2.0
chardet==5.2.0
structlog==25.4.0
